Add code coverage
